% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{dictionary_lookup}
\alias{dictionary_lookup}
\title{Look up column descriptions}
\usage{
dictionary_lookup(table = NULL, column = NULL)
}
\arguments{
\item{table}{Optionally filter to one table.}

\item{column}{Optionally filter to columns matching this text.}
}
\value{
A tibble of dictionary entries.
}
\description{
Look up column descriptions
}
\examples{
dictionary_lookup("pbp", "shot")
}
