% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_pbp}
\alias{load_pbp}
\alias{load_possessions}
\alias{load_lineup_stats}
\title{Load NBA play-by-play data}
\usage{
load_pbp(
  seasons = most_recent_season(),
  season_type = c("regular", "playoffs")
)

load_possessions(
  seasons = most_recent_season(),
  season_type = c("regular", "playoffs")
)

load_lineup_stats(
  seasons = most_recent_season(),
  season_type = c("regular", "playoffs")
)
}
\arguments{
\item{seasons}{Integer vector of seasons, labelled by the year the season
ends (e.g. 2025 for 2024-25).}

\item{season_type}{Either "regular" or "playoffs".}
}
\value{
A data frame of play-by-play events.
}
\description{
Load NBA play-by-play data
}
\examples{
\dontrun{
load_pbp(2025)
load_pbp(2023:2025, season_type = "playoffs")
}
}
