% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{most_recent_season}
\alias{most_recent_season}
\title{Most recent season with data available}
\usage{
most_recent_season()
}
\value{
An integer season, labelled by the year the season ends.
}
\description{
Most recent season with data available
}
